LayoutTests/media/modern-media-controls/scrubber-support/scrubber-support-drag-expected.txt

Testing the ScrubberSupport behavior when dragging the scrubber.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.currentTime is media.duration / 2
PASS media.currentTime is 0
PASS successfullyParsed is true

TEST COMPLETE
